Fix all resource resolution failures
